typescript: handle variables declared within for-loop initializer
Rather than specially handling for-loops, instead change around some
other code so our existing processing picks this up.

To understand why this works, consider:
  var x = 3;
  for (var y in ...)
The first is a VariableDeclarationStatement, while both both contain
a VariableDeclaration.

Previously we handled VariableDeclarationStatement specially to extract
defines/binding.  Instead if we handle VariableDeclaration, both forms
are covered.

So this fix for for loops instead only touches VariableDeclaration, and
then to delete a bit of our VariableDeclarationStatement handling.

Additionally, in writing the test for this change I found a bug in how
we handle the naming of the loop variables, so I fixed that too.