fix(typescript_indexer): log missing signature only if name is defined
We currently log an error message for nodes not included in a VName
signature but that have a "name" attribute. What we really care about,
though, is if the node has a name value; for instance, function
expressions like `(function())` have a name attribute, but it is always
`undefined`.

This reduces superfluous error logging.