feat(typescript_indexer): emit "aliases" edge on aliased types (#3870)
If a type is a singleton and not built-in, we can definitely emit an
"aliases" edge for it when it is aliased in a type definition, like
`type A = B`. Built-ins (like `number`) and unions are left untouched,
because they have zero and multiple derived VNames, respectively, so
there is no way to definitively point an edge toward those.