refactor(typescript_indexer): cleanup module path resolution (#3883)
We were handling module path resolution correctly previously, but doing
it in a confusing way. Since we only care about paths that define an
entire module, just query to see if the module reference is an entire
source file. If so, return the module path to that source file;
otherwise, we can bail because there is no unique path to reference.