feat: use bazel query to find targets to extract
Sometimes it may be easier to specify targets to build using a bazel query
rather than listing out explicit targets. Unfortunately, bazel cannot take many
arguments on the command line, so we need to shard our results across several
bazel invocations.

https://docs.bazel.build/versions/master/guide.html#target-patterns says that
target patterns should be prefixed by a " -- " so that negative target
patterns are not interpreted as a flag to bazel