refactor(typescript_indexer): emit "generates" edges for imports
Due to limitations in Kythe UIs, an anchor cannot be a definition and a
reference at the same time. For an import like

```typescript
import {foo} from './bar';
```

it would be nice to have local uses of `foo` "ref" `foo` in the import
statement and for `foo` in the import statement to "ref" the definition
of `foo` in './bar'. Since this is not possible, instead all local uses
of `foo` "ref" the definition of `foo` in './bar'.

The current way this is done is by having all Symbol VNames
corresponding to the external definition of `foo` also correspond to the
local definition of `foo`. This is somewhat unfavorable because it
breaks TypeScript semantics.

This PR introduces an alternate approach to acheiving the same UI xref
functionality by exploiting Kythe "generates" edges.

["generates"](https://kythe.io/docs/schema/#generates) edges connect two
semantic nodes (i.e. nodes that represent a language entity and not a
source code anchor) and are typically used for extralingual processes.
They have a nice property of folding cross-references for a UI; that is,
given a graph like
  (1) `A -ref-> D <-generates- B <-defines/binding- C`
a cross-reference between A and D is generated in a UI as if the graph
were
  (2) `A -ref-> B <-defines/binding- C`

For the TypeScript indexer, "D" in graph (1) is a local import VName.
Local usages can reference the local import, and such references will
later be folded to provide a UI xref from the local usage to the remote
definition.

Now that they are semantic nodes, local import symbols emit a
["name"](https://kythe.io/docs/schema/#name) node because they are
identifiers for a remote definition.

This commit introduces a change in the functionality of the indexer's
entry emissions but _not_ in its end result UI xrefs, so the change is
best classified as a refactoring.