feat(typescript_indexer): emit references to hops in imports
See longer discussion at #3934.

Due to limitations in Kythe UIs, an anchor cannot be a definition and a
reference at the same time. For an import like

```typescript
import {foo} from './bar';
```

it would be nice to have local uses of `foo` "ref" `foo` in the import
statement and for `foo` in the import statement to "ref" the definition
of `foo` in './bar'. Pending UI changes, instead all local uses of
`foo` "ref" the definition of `foo` in './bar'.

This PR introduces an alternate approach to acheiving more a
functionality more representative of TypeScript semantics by separating
local definitions and remote references in the same way the Python
indexer currently does (see #3934).

For a code like

```typescript
import {foo} from './bar'
```

The anchor at "import" now "defines/binding" the local definition of foo,
"LocalFoo". "LocalFoo"
- is a "variable" if it's a value, a "talias" if it's a type.
- has a subkind "import".
- "aliases" the remote definition of foo, "RemoteFoo".

The anchor at "foo" now "ref" the remote definition of foo, "RemoteFoo".

For a code like

```typescript
import {foo as baz} from './bar';
```

The functionality is more as expected. The anchor at "baz"
"defines/binding" the local definition of `baz`, and the anchor at "foo"
"ref" the remote definition of `foo`.

Finally, nodes are now emitted for both value and type imports. The
biggest difference is that local import declarations that are values
have a node kind "variable", and those that are types have a node kind
"talias". Import declarations that are both values and types have two
nodes emitted, one for each kind of import.