test(java_extractor): test/run Bazel Java extractor as aspect
Use a Bazel aspect to capture and transform a JavaInfo provider from a
Java compilation into a legacy JavaCompileInfo proto for consumption by
the existing Bazel Java extra action extractor.

Assuming the transform is correct, this provides the ability the test
the extractor within a Bazel build with the exact semantics of a build
with an attached extra action (without extra actions).