feature(typescript_indexer): emit reference for "this" keyword (#3973)
Emit a reference between "this" keywords that have a particular type and
the type of the object they refer to. For instance,

```typescript
//- @C defines/binding CType
class C {
  foo() {
    //- @this ref CType
    this.foo();
  }
}
```

Nothing is done for "this" keywords that do not refer to any particular
type. In

```typescript
let obj = {
  foo() {
    this.foo();
  }
}
```

`obj` has no particular type other than `object` and so no reference is
emitted. The same applies to class expressions.