fix(build): include C++14 options in default build flags
Fixes #4006. Recent versions of LLVM require some C++14 features, which may not
be enabled by default. Ensure Bazel enables them for local builds as well as
remote (CI) builds.

Also mention the need for C++14 in the setup instructions, to aid someone
trying to debug this from an old configuration (as I did).