feat(go extractor): add vnames.json support (#4069)
The new --rules flag allows providing a vnames.json file that re-maps package/file paths to custom corpus, root, path in the output.

The rules are applied to both packages and files. File paths are relative to the "root" directory, which is either GOPATH (if set) or the current working directory of the program.