typescript: index inputs that contain errors
This change makes us ignore TypeScript errors and index the parts
of the program that still have type information.  Previously we would
abort.

TypeScript in general has lots of accomodations around programs with
errors, so it's no trouble for TypeScript to do this.  (For example,
the compiler still wants to provide jump-to-definition features even
if you have a type error elsewhere in the same file.)

Within Google we don't run programs that have errors so code in this
state is rare, with one major exception: Clutz
(https://github.com/angular/clutz) output is invalid when its input
JavaScript is also invalid.  So this change has the main effect of
making Kythe able to index Clutz output.

We already had a test that exercised what happens when you have
bad input, so I adjusted that slightly to exercise this adjusted
logic.