fix(java extractor): Handle source files in jars
When we encounter a source file in a jar, don't try to add it as an
additional source path.

This includes a special case where the classpath is searched for sources
if no explicit sourcepath is given.

1) The URI generated by javac for files in a jar (JarFileObject) contains a
null path so the normal logic here will produce a null pointer exception.

2) The hooks into javac that we already have notice and record the source jar
usage, including the special case when it was specified as the classpath, not
the sourcepath.

Fixes #4186