fix(kzip info): don't count compilation units by corpus
the corpus specified in the unit vname is optional and only used as a
default for required_inputs that don't specify a corpus. instead of
keeping a count of the optionally-specified unit corpus, just count the
required_inputs and sources by corpus.

if a corpus can not be determined for a file, an error is recorded in
KzipInfo.critical_validation_errors