feat(jvm_indexer): allow JVM nodes to be refined by corpus/root/path (#4204)
It may be desirable to separate JVM nodes by corpus (and possibly
root/path).  This is important when mixing two similar corpora that
defines many of the same JVM entities that should not be considered
identical.  It is left up to the indexer whether to incorporate a
corpus/root/path into their emitted JVM nodes.