build proto metdata plugin from source in release
The c++ code generated by the plugin #includes some headers from the
protobuf "runtime", which is required to be of the same protobuf version
as the generated code. Previously, we shipped the plugin in binary form,
which required clients to use the same version of protobuf as our
release was compiled at. This version constraint ended up being a
hassle, so we're shipping the plugin in source form instead so it will
be built against the same version of protobuf as the client project.

This unfortunately requires editing the client WORKSPACE file to add the
kythe_release repository and run an included bazel macro for retrieving
dependencies.

```
local_repository(
    name="kythe_release",
    path="/path/to/kythe/kythe-v0.0.35",
)

load("@kythe_release//:external.bzl", "kythe_release_dependencies")
kythe_release_dependencies()
```