chore(java indexer): log filemanager exceptions
The filemanager sometimes logs errors to stderr without any context and that makes it more difficult than it should be to debug problems. Add an explicit log invocation in all methods called via reflection. We could add the log to non-reflective calls as well, but that would be a more invasive change.