feat(java indexer): better support system modules
The Java libraries handle the system module path in a special way so our normal
methods for reading from a compilation unit instead of the local file system do
not work. The easiest way to deal with this is to treat system modules
specially in our indexer and unpack them to a local temp directory for java to
read from. This is not ideal because it pollutes the local file system but given
what the Java libraries do, it doesn't look like there are any better options.

H/T @shahms for the majority of this PR.