refactor(kzip info): allow composing other operations in kzip.Scan (#4220)
Scanning a kzip can be slow for large files, so if multiple operations
need to be done, it's preferable that they all be done in one
kzip.Scan() instead of separately. This change moves most kzip info
logic into an "Accumulator" class that can be called from your own
invocation of kzip.Scan().

The motivation for this is that I need to get summary info for a kzip
and also search for certain units by file path. Doing these in one
kzip.Scan() saves a lot of time.