fix(java indexer): Clean up after exceptions
We may store some temporary files on disk when we create our file manager.
Normally we clean them up when the compilation details object is closed, but if
an exception happens before we return the compilations details object, we will
leave artifacts on disk. Handle exceptions while we are creating the
compilation details object so this doesn't happen.