chore(doc): remove stale script
The extract.sh script is meant to be run in a docker container. The
instructions on our examples page no longer work, so remove it. There is
still a reference to this script later in the page, but that is a more
invasive change and will require more work to update.