fix(textproto): support direct instantiation of protobuf.Any messages (#4259)
Textproto usage of Any messages comes in two forms. You can specify the
Any directly via the `type_url` and `value` fields or you can specify the
message as a literal. If AnalyzeAnyTypeUrl() is unable to find a literal
starting with a type url enclosed in brackets, it returns an error to
incicate it's a directly-specified Any.