fix(java): disable errorprone plugin for java extraction
The errorprone plugin complicates the build due to certain other
flags it requires (such as -XDcompilePolicy=byfile) and is not
necessary for extraction.

Another workaround would be to disable the errorprone plugin in the
project's gradle config, but that's often a fairly invasive change
(AFAIK) and requires removing config options in several places.