fix(bazel_go_extractor): record canonical importpath for archives (#4425)
This should fix cross-module linking for go code extracted through bazel.

Go package dependencies are pulled in as static archives (.a files) and are mentioned in the arguments like:

-arc "github.com/google/subcommands=github.com/google/subcommands=bazel-out/k8-fastbuild/bin/external/com_github_google_subcommands/linux_amd64_stripped/go_default_library%/github.com/google/subcommands.a="

This change parses the -arc arguments and builds a map of filepath -> importpath, which is then used to attach GoPackageInfo details messages to the corresponding required_inputs in the compilation unit. The indexer later picks these up and refers to package dependencies by the correct vname/importpath.