Support deprecated tags in TS marked source
While reading the [TS 4.0 release notes](https://devblogs.microsoft.com/typescript/announcing-typescript-4-0-beta/),
I realized that the TS indexer currently doesn't tag entities that are
marked as deprecated. There's a formal API for this in TS 4.0, but we
accurately determine whether a node is deprecated (and why) by scanning
the JSDoc tags of that node.