feat(rust): Create KytheWriter component for Rust indexer (#4565)
This PR adds a new component to the Rust indexer library. The new writer component contains a `KytheWriter` trait, which is implemented by the new `CodedOutputStreamWriter`. The `KytheWriter` trait will be used by the indexer to write Kythe graph entries to some output during indexing. The `CodedOutputStreamWriter` works by taking an object that implements `Write` and creating a protobuf `CodedOutputStream` that will output to that writer. When an `Entry` is written, it's length is first written using a varint32, followed by the actual `Entry` object, per the Kythe spec. The `ArrayWriter` implemented in the integration tests works by appending Entry objects to a `Vec`.