Do not group files in nested test directories by default (#4591)
Previously test files for the TS indexer that were in a common
subdirectory would be tested together. This is useful in a lot of cases
but isn't the behavior we want for marked source tests, which are put in
a separate directory because they test a lot of code that is specific to
marked source and whose files are independent from each other.

To ensure that files for marked source tests are treated independently,
this commit groups tests only when the nested test directory ends with
`_group`.