feat(rust): Unify Kythe Rust dependencies (#4611)
This PR places all of the Kythe Rust dependencies in one directory: `//kythe/rust/cargo`. Previously, each component had its own directory that contained the component's dependencies. However, this quickly became a messy system, leading to file bloat and dependency duplication. This folder can only hold one version of each dependency, though. If a situation arises where a Kythe crate needs a different version of a dependency than what is in this directory, that version can be placed in the crate's directory instead.