Works around the -fPIC issue for rust
Rust binaries are compiled with -fPIC by default, and this fights with
bazel's default of -fno-pic when using `-c opt` mode.  This has caused
the compilation to fail.

After some running around with hair on fire, I found
https://github.com/bazelbuild/rules_rust/issues/118, which seems to
work around the issue.

I think in the long run, rules_rust should handle this automagically for
the users.