KZIP Writer C API (and Rust FFI) (#4608)
* feat(rust): Adds a C API for the kzip writer

This will allow me to reuse the C++ kzip writing code for
the rust indexer instead of redoing that work in rust.

Issue #4606

* Renames the 'status' library

Somewhat surprisingly -- if there are two libraries in the global namespace
that have the same name, bazel does not disambiguate them, and an attempt
to link them will fail.  Not sure why, but that's what happens.

* fixup: handled review comments

Simplified and cleared up issues raised during
code review.

* Works around the -fPIC issue for rust

Rust binaries are compiled with -fPIC by default, and this fights with
bazel's default of -fno-pic when using `-c opt` mode.  This has caused
the compilation to fail.

After some running around with hair on fire, I found
https://github.com/bazelbuild/rules_rust/issues/118, which seems to
work around the issue.

I think in the long run, rules_rust should handle this automagically for
the users.

* Document no NUL byte at the end of results