feat(rust): Add Bazel rule for running Rust indexer tests (#4612)
This PR creates a Bazel rule for running tests on the Rust indexer using the Kythe verifier. It is based on the existing `go_indexer_test` rule, and utilizes the `go_verifier_test` macro. This PR also includes a small Rust file to sanity-test the rule until the indexer starts emitting data.