fix(rust): Ensure that Rust extractor creates a top-level folder (#4616)
This PR fixes a bug in the Rust extractor where a top-level folder was not being created in the kzips, causing them to be invalid. The extractor will now create a top-level folder named "root". The indexer has been fixed to properly detect the root folder name. 

This PR adds an exclusion for `kythe/java/com/google/devtools/kythe/platform/java/helpers/SignatureGenerator.java` from the arcanist spellchecker. This prevents the linter from failing due to a false-positive.