feat(fuchsia): Fuchsia-specific extractor
This commit implements a compilation extractor for Fuchsia OS's Rust
source code. Please refer to the README.md file for a run-down
of the internal workings of the extractor.

While the extractor is able to extract compilation into kzip files
as is, I expect that there may be touch-ups needed if we discover
integration issues.

Closes #4606