chore(proto_indexer): remove unused FileVNameGenerator (#4627)
The FileVNameGenerator only produces useful VName lookups if given a
vname config file or if other defaults are set. In this case it was just
being used empty.

Because the proto indexer only needs file vnames for .proto files and
these all have vnames specified in the compilation unit, there isn't
need for other configuration.

In the case that a proto is referenced, but not in the compilation unit,
an empty vname would be used. This shouldn't happen though because the
proto parser itself would fail and the indexing code wouldn't get run.

We could consider changing the lookup functions to return a StatusOr to
handle that case better.