feat(fuchsia): Fuchsia-specific extractor (#4624)
* Adds more required Rust crates

The crates are not yet used, but are added in a separate commit
to keep the crate addition noise out of the code review.

A followup commit will add the usage

* feat(fuchsia): Fuchsia-specific extractor

This commit implements a compilation extractor for Fuchsia OS's Rust
source code. Please refer to the README.md file for a run-down
of the internal workings of the extractor.

While the extractor is able to extract compilation into kzip files
as is, I expect that there may be touch-ups needed if we discover
integration issues.

Closes #4606

* Fixed up buildbot errors

* Update README.md

Adds a short explanation for why the separation between
the build system and the extractor is not as clean as
it could be in theory.

While not ideal today, I think this is will improve
significantly in the medium term.

* Fixed up the README.md file

Based on the code review comments.

* Some more code review fixups

Wording in README around the 'root' labels, and
code change to use an empty label '' for source
files from the tree.