fix(rust): parallelize the extraction process
The Fuchsia compilation extraction is easy to parallelize, so
we do so.  The rayon library farms out the extraction process,
and the extraction workers won't interfere with each other as
they only read inputs, and are guaranteed not to step over
each other.

This should speed up the extraction significantly.