fix(rust): Fixes the kzip generation
The old code filled FileInfo.path with a path relative to the corpus
root, but should instead have filled it with a path relative to the
build directory.

Added a regression tests, and added *most* of an integration test
which runs kzip info... for additional verification of adherence
to the spec.

Howeer, that last test is currently not enabled as it is apparently
not allowed to subprocess to a binary.