fix(rust): Handles different roots for Fuchsia (#4653)
Adds sorting code that bins all the Fuchsia source
files into separate roots, based on the relative
path of a file with respect to the directory the
root builds in.

There are about a dozen different roots, some of which
are "" (the "default" source, which is the "in-tree" code),
the "gen" (source code generated by build rules), "fidling_gen"
(the source code generated by the FIDL compiler) and so on.

Regression test added to confirm that the paths are set as
expected.

Also, added the crate `serial_test` and the `#[serial]` annotations to
serialize test runs.  It seems that `TempDir` somehow creates temporary
directories with conflicting suffixes, which then interfere with each
other, causing flaky test behavior.  With this change, flakiness went
from 1/1000 to 0/30000.

Flakiness tested by:

```bash
bazel test //kythe/rust/fuchsia_extractor/... --runs_per_test=30000
```