feat(textproto): prototype support for indexer plugins (#4623)
* plugins are linked into the indexer binary
* after a textproto is parsed, a list of enabled plugins is determined based on the top-level proto message name. plugins are constructed once per textproto file.
* each time a string field value is seen in the input textproto, the plugin's AnalyzeStringField() method is called
the plugin is given the text span(s) of the string value. it can decide to do nothing or emit 1+ anchors covering subsets of the span(s)
* a subset of the TextprotoAnalyzer class is exposed through the PluginApi interface, which the plugin can use to add kythe entries and get info about the textproto

Caveats:
This will only support "simple" plugins that don't need to read other textprotos as dependencies. The plugin only has access to the current textproto and any .protos it depends on. The plugin is free to link in whatever data it wants at build time, but can not dynamically load other deps at index time (since they won't be in the compilation unit). This plugin support should work for several proposed/existing textproto-based features/languages in google3, but not all.