fix(bazel): align options between `query` and `build`
I noticed that bazel often re-analyzes a significant number of dependencies,
particularly when using ibazel. I'm able to reproduce the issue by doing a bazel
test/build, then running a quuery. When going back and forth between the two,
bazel re-fetches a lot of dependencies and re-analyzes them. This is
especially bad when using ibazel because it runs a bazel query at startup and
whenever the build graph changes.