chore(typescript_indexer): group related test files in subdirectories. (#4719)
Tests that verify indexing of importing/exporting symbols require indexing two or more files at the same time. Currently some of these tests live in the root testdata directory which means that each file is indexed and verified individually defeating the purpose of the test. Moving them to separate _group directories solves that.

Co-authored-by: Shahms King <shahms@google.com>