feat(verifier): compile and run a simple datalog program
This PR extracts code from Souffle that transforms an input
program string into an optimized representation. It was necessary
to do this because the code in question is not exported as
a library but is instead inlined in main(). It then checks that
running a simple example program produces the correct results.