feat(verifier): build a souffle frontend that can read an entrystream
This PR builds a version of the souffle frontend that can read an
entrystream directly. For example, the program

```
.type vname = [
  signature:symbol,
  corpus:symbol,
  path:symbol,
  root:symbol,
  language:symbol
]

.decl entry(source:vname, kind:symbol, target:vname, value:symbol)
.input entry(IO=kythe_entrystream)

.decl outentry(source:vname, kind:symbol, target:vname, value:symbol)
.output outentry(IO=json)

outentry(s,k,t,v) :- entry(s,k,t,v).
```

when invoked as
```
./bazel-bin/kythe/cxx/verifier/souffle_with_kythe example.dl \
    < in.entries
```

will print the entries as JSON.

libffi still doesn't work on Darwin and the fix isn't immediately
clear (I did look at a different BUILD file, but I think their
version of libffi was different; ours seems to be missing some
Darwin headers for x86_64?). I've stubbed the relevant symbols
out to make the linker happy.

I don't know if this is the most efficient representation of the
Kythe graph or if it has other problems, so please don't rely on
it (and suggestions for improvement are welcome).