fix(cxx_indexer): refactor lvalue handling for dataflow (#4907)
* fix(cxx_indexer): refactor lvalue handling for dataflow

Tag particular AST nodes as writes. This is cleaner than the previous
approach, which tries to scramble up the tree to figure out if a node
is in a write context. This PR also unifies lvalue interpretation
and influence target identification.

We can probably turn dataflow edges on by default, but that will be
a separate PR.