Apply semantic metadata rules for function calls for #5037
This PR adds a cache of alternate semantics that maps from
source locations to alternate semantics as specified in metadata
files. When a relevant declaration is discovered at one
of these source locations, its canonical representative is
added to a separate cache so it can be looked up directly.
This way, we don't have to complicate IndexedParentMap any more
(or worse, add another prepass over the AST).