feat(cxx_indexer): use default corpus for meta, builtin nodes (#5008)
This should be a no-op unless --use_compilation_corpus_as_default is
enabled, in which case these nodes will be assigned the compilation
unit's corpus.

KytheGraphObserver's constructor emits nodes for builtins and meta
nodes, so setting the default corpus later means it won't apply to these
nodes. The constructor has been modified to accept a default_corpus
parameter, which is set to the compilation unit's corpus when
--use_compilation_corpus_as_default is enabled.

A new test was added that reads the indexer output for a test example
and checks that each vname has a non-empty corpus.