feat(cxx_indexer): Allow VNames as NodeIds for #5037 (#5041)
* feat(cxx_indexer): Allow VNames as NodeIds for #5037

This PR allows VNames to be smuggled inside NodeIds. This is necessary
because the alternate semantic metadata rules need to be applied
inside IndexerASTHooks which only deals in NodeIds.

This implementation seemed slightly less evil than encoding a pointer
to a VName in a string. I considered using proto wire format, but
the lifetimes involved make things hard (the indexer likes to deal
in VNameRefs for efficiency (without them IIRC we spent a significant
percentage of cycles on copying strings around) and there's nowhere
for a temporary VName to go).