fix: the mac build
Macs are (mostly) case-insensitive. A change to LLVM resulted in
a file called CXXABI.h being visible in the system search path
ahead of the actual cxxabi.h. This caused certain Abseil libraries
to fail to build as they depend on the content of that header.

This PR pushes Abseil to head (why not?), LLVM to head (as they
fixed this problem upstream), and rules_cc to head (because the
LLVM fix requires a more recent version).

Addresses #5133