fix(cxx_indexer): remove another source of corpusless nodes
In some instances, a c++ symbol/node is not associated with a file and ends up with an empty corpus. This seems to happen when a macro is defined on the command line (i.e. "-DMY_MACRO=1"), then later referred to in the source code. This change sets the corpus to the default value (from the compilation unit) instead.

TODO: test this - it works in google3 localrun, but getting the issue to reproduce in the oss project is tricky