fix(cxx_indexer): remove another source of corpusless nodes (#5247)
* fix(cxx_indexer): remove another source of corpusless nodes

In some instances, a c++ symbol/node is not associated with a file and ends up with an empty corpus. This seems to happen when a macro is defined on the command line (i.e. "-DMY_MACRO=1"), then later referred to in the source code. Interestingly it's not enough to just define the macro on the command line. You have to define it, undefine it, then define it again... Maybe this test case can be simplified and it's likely there's a deeper underlying bug here.

This change sets the corpus to the default value (from the compilation unit) instead.