fix(kzip_info): don't count c++ stdlib includes as invalid absolute paths
in general, paths in the kzip should be relative (to the root of the repo), so we count absolute paths as errors. in some cases like c++ stdlib includes, these are expected and should not count as errors. There are likely other exceptions to the rule that we'll want to add in the future