feat(go_indexer): rework --use_compilation_corpus_as_default (#5338)
Most of this change is a flag/option rename. The most important bit is in rewrittenCorpusForVName() in emit.go.

The --override_stdlib_corpus wasn't working as intended in google3 because it identifies stdlib nodes by the 'golang.org' corpus, but in google3, these are assigned the 'google3/golib' corpus. I renamed --use_compilation_corpus_as_default to --use_compilation_corpus_for_all to better indicate it's function, which is to change the corpus of all VNames (including stdlib and builtin items) at emit time to that of the compilation corpus, which is what we want for oss repos. rewrittenCorpusForVName() has been changed so that when --use_compilation_corpus_for_all is disabled, the compilation unit's corpus is only used for items that don't already have a corpus set, which is what we want for google3 and google3/golib. Builtin types such as string or bool will still be assigned 'golang.org', which is what we were already doing before I started messing with corpus rewriting. We could use --override_stdlib_corpus to change these to 'golang.org', but I'm inclined to remove the override flag and just continue using 'golang.org' for builtins.