chore(typescript_indexer): consolidate indexer options in a single object
This is necessary to add more options in future. Currently each options requires
new method argument and that hurts readability.

Also using this opportunity modify a bit API of index() function and plugins. Now we pass
files to index as explicit `paths` array rather than it being part of IndexerHost.
That way index methods are a bit easier to understand of *what* exactly they need to
index.