feat(typescript): merge constructor nodes (#5609)
This PR changes what nodes are emitted for TS classes. Before the change for the following class:

```
class Foo {
  constructor() {}
}
```

there were 3 nodes:

* `Foo#type` - node representing type of the class.
* `Foo` - node representing value of the class.
* `Foo#type.constructor` - node representing constructor (if class an explicit constructor).

This PR merges the last two nodes into Foo. So after this change it will look like this:

* `Foo#type` - node representing class.
* `Foo` - node representing constructor.

Anchor for the constructor node is chosen based on whether class has an explicit constructor.

With the new change clicking on constructor in the class declaration users will see all instantiations of the class.