chore(typescript_indexer): extract plugin API to a separate file (#5684)
Inputs such as CompilationUnit, IndexerHost and others as well as Plugin interface are moved to `plugin_api.ts`.

This way it's easier for plugin developers to see what API is and what they have access to. Additionally it allows for further splitting indexer.ts into sub files as currently that file is almost 3k lines.

Additionally reworks indexer.ts to introduce TypescriptIndexer plugin which is a main plugin that indexes TS files. By doing that we make sure that Plugin interface provides enough capabilities for proper indexing.

This PR will require a bunch of changes in google3 as other TS plugins import interfaces that were moved to `plugin_api.ts`. But it should be trivial to update imports to point at the new file.